centralize subscriber setup logic from queries
